% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_neighbours.R
\name{find_neighbors2}
\alias{find_neighbors2}
\title{find_neighbors2.}
\usage{
find_neighbors2(sce, platform)
}
\arguments{
\item{sce}{is a SingleCellExperiment object containing PCA and position informatin.}

\item{platform}{is the name of spatial transcriptomic platform. Specify 'Visium' for hex lattice geometry or 'ST' for square lattice geometry. Specifying this parameter is optional as this information is included in their metadata.}
}
\value{
a sparse matrix recording the information of neighborhood.
}
\description{
find_neighbors2 was used to find the neighborhood of spot.
}
\details{
find_neighbors2 was used to find the neighborhood of spot.
}
\examples{
data(sce)
platform = "ST"
Adj <- find_neighbors2(sce, platform)
}
