% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-MEB.R
\name{parafun}
\alias{parafun}
\title{parafun.}
\usage{
parafun(
  y,
  Adj,
  G,
  beta_grid = seq(0, 4, 0.2),
  PX = TRUE,
  maxIter_ICM = 10,
  maxIter = 50
)
}
\arguments{
\item{y}{is n-by-d PCs.}

\item{G}{is an integer specifying the numbers of mixture components (clusters) for which the BIC is to be calculated. The default is K = 2:9.}

\item{maxIter_ICM}{is the maximum iteration of ICM algorithm. The default is 10.}

\item{maxIter}{is the maximum iteration of EM algorithm. The default is 50.}

\item{Adj_sp}{is a sparse matrix of neighborhood.}

\item{platform}{is the name of spatial transcriptomic platform. Specify 'Visium' for hex lattice geometry or 'ST' for square lattice geometry. Specifying this parameter is optional as this information is included in their metadata.}

\item{bet_grid}{is a numeric vector specifying the smoothness parameter of Random Markov Field. The default is seq(0,4,0.2).}
}
\value{
a list, We briefly explain the output of the SC.MEB. 
 
The item 'x' storing clustering results.

The item 'BIC' contains all BIC value. 

The item 'ell' is the opposite log-likelihood. 

The item 'mu' is the mean of each component.

The item 'sigma' is the variance of each component.
}
\description{
parafun implements the model SC-MEB for fixed number of clusters
}
\details{
SC.MEB implements the model SC-MEB, spatial clustering with hidden Markov random field using empirical Bayes.
}
