% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PC}
\alias{PC}
\title{simulated PCs}
\format{
It is a matrix containing 5 PCs

the variables are listed as following
\describe{
  \item{PC1}{The 1th PC}
  \item{PC2}{The 2th PC}
  ...
  \item{PC5}{The 5th PC}
}
}
\usage{
data(PC)
}
\description{
A dataset containing PCs
}
\examples{
## run the PC with the Gaussian mixture model
data(PC)
out1 = mclust::Mclust(PC,G = 2)
}
\keyword{datasets}
