% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-MEB.R
\name{SC.MEB}
\alias{SC.MEB}
\title{SC.MEB.}
\usage{
SC.MEB(
  y,
  Adj_sp,
  beta_grid = seq(0, 4, 0.2),
  K_set = 2:10,
  parallel = TRUE,
  num_core = 5,
  PX = TRUE,
  maxIter_ICM = 10,
  maxIter = 50,
  ...
)
}
\arguments{
\item{y}{is n-by-d PCs.}

\item{Adj_sp}{is a sparse matrix of neighborhood. It is often generated from function find_neighbors2 or getneighborhood_fast.}

\item{beta_grid}{is a numeric vector specifying the smoothness parameter of Random Markov Field. The default is seq(0,4,0.2).}

\item{K_set}{is an integer vector specifying the numbers of mixture components (clusters) for which the BIC is to be calculated. The default is K = 2:9.}

\item{parallel}{is a logical value to decide whether the function SC.MEB run in parallel. The default is True.}

\item{num_core}{is an integer value to decide how many cores are used to run SC.MEB in parallel.}

\item{PX}{is a logical value to decide whether to use parameter expansion in EM algorithm}

\item{maxIter_ICM}{is the maximum iteration of ICM algorithm. The default is 10.}

\item{maxIter}{is the maximum iteration of EM algorithm. The default is 50.}

\item{platform}{is the name of spatial transcriptomic platform. Specify 'Visium' for hex lattice geometry or 'ST' for square lattice geometry. Specifying this parameter is optional as this information is included in their metadata.}
}
\value{
a list, We briefly explain the output of the SC.MEB. 
 
The item 'x' is 3-dimensional n-by-b-by-q matrix, storing all clustering results for each K and beta. n is the number of cells, b is the length of vector 'bet', q is the length of vector 'K'. 

The item 'BIC' contains all BIC value for each K and beta. 

The item 'ell' is the opposite log-likelihood for each beta and K. 

The item 'mu' is the mean of each component for each beta and K.

The item 'sigma' is the variance of each component for each beta and K.
}
\description{
SC.MEB implements the model SC-MEB, spatial clustering with hidden Markov random field using empirical Bayes.
}
\details{
SC.MEB can implements the model SC-MEB in parallel which can improve the speed of the computation.
}
\references{
Yang Y, Shi X, Zhou Q, et al. SC-MEB: spatial clustering with hidden Markov random field using empirical Bayes[J]. bioRxiv, 2021.
}
