% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce}
\alias{sce}
\title{A simulated SingleCellExperiment}
\format{
It is a SingleCellExperiment object with gene expression and meta information
}
\usage{
data(sce)
}
\description{
A dataset of SingleCellExperiment
}
\examples{
## find the neighborhood of spots in SingleCellExperiment
data(sce)
out = find_neighbors2(sce, "ST")
}
\references{
Amezquita R A, Lun A T L, Becht E, et al. Orchestrating single-cell analysis with Bioconductor[J]. Nature methods, 2020, 17(2): 137-145.
}
\keyword{datasets}
