% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-MEB.R
\name{selectK}
\alias{selectK}
\title{selectK.}
\usage{
selectK(SCobject, K_set = 2:10, criterion = "BIC", pen.const = 1)
}
\arguments{
\item{SCobject}{is an object generated from SC.MEB function.}

\item{K_set}{is a integer vector used in SC.MEB. The default is 2:10}

\item{criterion}{is a character specifying the criterion for selecting K. The default value is BIC. The alternative value MBIC can also be used.}
}
\value{
a list contains two items. one is for the best K and the other corresponds to clustering labels of n spots.
}
\description{
The function selectK is used to select the best K according to BIC or Modified BIC criterion.
}
\details{
The function selectK is used to select the best K according to BIC or Modified BIC criterion.
}
