% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-MEB.R
\name{ClusterPlot}
\alias{ClusterPlot}
\title{ClusterPlot.}
\usage{
ClusterPlot(out, pos, size = 5, shape = 15)
}
\arguments{
\item{out}{is the output of function selectK.}

\item{pos}{is a n-by-2 matrix of position.}

\item{size}{is a positive value for characterizing the size of point in the plot, which is the same as size in ggplot2.}

\item{shape}{is a positive value for characterizing the shape of point in the plot, which is the same as shape in ggplot2.}
}
\value{
a ggplot2 object.
}
\description{
The function ClusterPlot is used to Visualize spatial clusters.
}
\details{
The function ClusterPlot is used to Visualize spatial clusters.
}
\examples{
pos = cbind(rep(1:5, each=5), rep(1:5, 5))
out = list()
out[[1]] = ""
out[[2]] = rep(1:5, each = 5)
ClusterPlot(out, pos)
}
