% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SC-MEB.R
\name{selectKPlot}
\alias{selectKPlot}
\title{selectKPlot.}
\usage{
selectKPlot(SCobject, K_set = 2:10, criterion = "BIC", pen.const = 1)
}
\arguments{
\item{SCobject}{is a object generated from SC.MEB function.}

\item{K_set}{is the corresponding K_set used in your previous function SC.MEB.}

\item{criterion}{is a character specifying the criterion for selecting K. The default is BIC, the alternative criterion MBIC can also be used.}
}
\description{
The function selectKPlot is used to demonstrate the scatter plot of BIC or Modified BIC vs K for selecting the best K.
}
\details{
The function selectKPlot is used to demonstrate the scatter plot of BIC or Modified BIC vs K for selecting the best K.
}
